refactor(accounts): split persistence workflows
