+++ src/features/AccountManagement/components/AccountDialog/hooks/useAccountDialog.ts
@@ -37,16 +37,14 @@ import {
   type ResolveAccountBrowserSessionOptions,
 } from "~/services/accountBrowserSession"
 import { autoDetectAccount } from "~/services/accounts/accountAutoDetection"
+import { validateAndSaveAccount } from "~/services/accounts/accountCreation"
 import { findExactCredentialDuplicateAccountId } from "~/services/accounts/accountDedupe"
 import {
   isValidAccount,
   parseManualQuotaFromUsd,
 } from "~/services/accounts/accountFormValidation"
 import { normalizeAccountIdentity } from "~/services/accounts/accountIdentity"
-import {
-  validateAndSaveAccount,
-  validateAndUpdateAccount,
-} from "~/services/accounts/accountPersistence"
+import { ACCOUNT_SAVE_FEEDBACK_LEVELS } from "~/services/accounts/accountPersistence/constants"
 import {
   ACCOUNT_POST_SAVE_WORKFLOW_STEPS,
   ACCOUNT_TOKEN_INVENTORY_STATE_KINDS,
@@ -58,6 +56,7 @@ import {
 } from "~/services/accounts/accountPostSaveWorkflow"
 import { doAccountSiteIdentitiesMatch } from "~/services/accounts/accountSiteProfile"
 import { accountStorage } from "~/services/accounts/accountStorage"
+import { validateAndUpdateAccount } from "~/services/accounts/accountUpdate"
 import type { CreatedRuntimeSecret } from "~/services/accounts/createdRuntimeSecret"
 import { getSiteName } from "~/services/accounts/siteName"
 import {
@@ -2503,7 +2502,7 @@ export function useAccountDialog({
                 name: siteName,
               })
 
-      if (result.feedbackLevel === "warning") {
+      if (result.feedbackLevel === ACCOUNT_SAVE_FEEDBACK_LEVELS.Warning) {
         const warningAccountId =
           typeof result.accountId === "string" && result.accountId.trim()
             ? result.accountId.trim()
