+++ tests/features/AccountManagement/hooks/useAccountDialog.currentTabDetection.test.tsx
@@ -77,11 +77,9 @@ vi.mock("~/services/accounts/siteName", () => ({
   getSiteName: mockGetSiteName,
 }))
 
-vi.mock("~/services/accounts/accountPersistence", async (importOriginal) => {
+vi.mock("~/services/accounts/accountCreation", async (importOriginal) => {
   const actual =
-    await importOriginal<
-      typeof import("~/services/accounts/accountPersistence")
-    >()
+    await importOriginal<typeof import("~/services/accounts/accountCreation")>()
   return {
     ...actual,
     validateAndSaveAccount: mockValidateAndSaveAccount,
