+++ tests/services/accountPersistence.saveAndUpdate.test.ts
@@ -1,12 +1,10 @@
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest"
 
 import { SITE_TYPES } from "~/constants/siteType"
-import {
-  MANUAL_ADD_ACCOUNT_DATA_FETCH_TIMEOUT_MS,
-  validateAndSaveAccount,
-  validateAndUpdateAccount,
-} from "~/services/accounts/accountPersistence"
+import { validateAndSaveAccount } from "~/services/accounts/accountCreation"
+import { MANUAL_ADD_ACCOUNT_DATA_FETCH_TIMEOUT_MS } from "~/services/accounts/accountCreationTimeout"
 import { accountStorage } from "~/services/accounts/accountStorage"
+import { validateAndUpdateAccount } from "~/services/accounts/accountUpdate"
 import { OpenRouterManagementKeyRequiredError } from "~/services/apiService/openrouter/errors"
 import { API_ERROR_CODES, ApiError } from "~/services/apiTransport/errors"
 import {
